Create a dedicated `SymbolDef::Variant` to fix finding references of enum variants
Previously, `SymbolDef::Item` was used for enum variants. `ItemDef::name` returned the name of a **lookup item** related to the item, which for enum variants was **the enum**, not the variant itself. This caused a bug because usage finding uses the symbol's name as a text search needle, which made the algorithm attempt looking for references to the enum, yielding effectively no usage results.

fix #129 - this was the last FIXME comment related to that issue that was left, even though the bug eventually appeared to be completely unrelated.

commit-id:374fbe2d